Refactor: Header.styles 컴포넌트 분리, 적용
* style 파일 이름 변경
