+++ src/pages/tvShow/on_air.tsx
@@ -1,4 +1,4 @@
-import { StyledWhiteBody } from '@/styles/pages/Category';
+import { StyledWhiteBody } from '@/styles/pages/Category.styles';
 
 import SEO from '@/components/SEO';
 import HeaderContainer from '@/containers/HeaderContainer';
@@ -15,6 +15,7 @@ export default function OnAir() {
           <h3>
             <span>TV 방영 중</span>
           </h3>
+
           <section>
             <TVContainer request='on_air' layout='wrap' />
           </section>
