+++ src/styles/pages/Category.styles.ts
