+++ src/styles/pages/Search.styles.ts
